Moved config.py into entropylab dir to be packaged
